Add update-profile for actors, fix typo
